pyproject.toml
[project]
name = "event-board-itorum"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.10"
dependencies = [
    "celery[beat]>=5.5.3",
    "django>=5.2.2",
    "django-celery-beat>=2.8.1",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "djangorestframework>=3.16.0",
    "djangorestframework-simplejwt>=5.5.0",
    "dotenv>=0.9.9",
    "drf-yasg>=1.21.10",
    "isort>=6.0.1",
    "psycopg2-binary>=2.9.10",
    "pytest>=8.4.0",
    "pytest-cov>=6.2.1",
    "pytest-django>=4.11.1",
    "redis>=6.2.0",
    "ruff>=0.11.13",
]
